requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
geographiclib>=2.0
pydantic>=2.8.0
python-multipart>=0.0.6
aiohttp>=3.8.0
requests>=2.28.0
openai>=1.0.0
python-dotenv>=1.0.0
# Langchain相关依赖
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0